Move to impls to have higher priority over the builtin write extension traits
