--- packages/stores/src/builtin/mod.rs
@@ -1,14 +1,0 @@
-mod hashmap;
-mod vec;
-pub use vec::*;
-mod slice;
-pub use hashmap::*;
-pub use slice::*;
-mod result;
-pub use result::*;
-mod option;
-pub use option::*;
-mod deref;
-pub use deref::*;
-mod index;
-pub use index::*;
