+++ packages/stores/src/impls/vec.rs
@@ -1,22 +1,7 @@
 use crate::store::Store;
 use dioxus_signals::Writable;
 
-mod private {
-    pub trait Sealed {}
-}
-
-/// A trait for `Store` that provides methods for working with `Vec` types.
-///
-/// # Example
-/// ```rust, no_run
-/// use dioxus_stores::*;
-/// let mut store = use_store(|| vec![1, 2, 3]);
-/// store.push(4);
-/// ```
-pub trait VecStoreExt: private::Sealed {
-    /// The item type of the vector.
-    type Item;
-
+impl<W: Writable<Target = Vec<T>> + Copy + 'static, T: 'static> Store<Vec<T>, W> {
     /// Pushes an item to the end of the vector. This will only mark the length of the vector as dirty.
     ///
     /// # Example
@@ -25,7 +10,10 @@ pub trait VecStoreExt: private::Sealed {
     /// let mut store = use_store(|| vec![1, 2, 3]);
     /// store.push(4);
     /// ```
-    fn push(&mut self, value: Self::Item);
+    pub fn push(&mut self, value: T) {
+        self.selector().mark_dirty_shallow();
+        self.selector().write_untracked().push(value);
+    }
 
     /// Removes an item from the vector at the specified index and returns it. This will mark items after
     /// the index and the length of the vector as dirty.
@@ -37,7 +25,11 @@ pub trait VecStoreExt: private::Sealed {
     /// let removed = store.remove(1);
     /// assert_eq!(removed, 2);
     /// ```
-    fn remove(&mut self, index: usize) -> Self::Item;
+    pub fn remove(&mut self, index: usize) -> T {
+        self.selector().mark_dirty_shallow();
+        self.selector().mark_dirty_at_and_after_index(index);
+        self.selector().write_untracked().remove(index)
+    }
 
     /// Inserts an item at the specified index in the vector. This will mark items at and after the index
     /// and the length of the vector as dirty.
@@ -48,7 +40,11 @@ pub trait VecStoreExt: private::Sealed {
     /// let mut store = use_store(|| vec![1, 2, 3]);
     /// store.insert(1, 4);
     /// ```
-    fn insert(&mut self, index: usize, value: Self::Item);
+    pub fn insert(&mut self, index: usize, value: T) {
+        self.selector().mark_dirty_shallow();
+        self.selector().mark_dirty_at_and_after_index(index);
+        self.selector().write_untracked().insert(index, value);
+    }
 
     /// Clears the vector, marking it as dirty.
     ///
@@ -58,7 +54,10 @@ pub trait VecStoreExt: private::Sealed {
     /// let mut store = use_store(|| vec![1, 2, 3]);
     /// store.clear();
     /// ```
-    fn clear(&mut self);
+    pub fn clear(&mut self) {
+        self.selector().mark_dirty();
+        self.selector().write_untracked().clear();
+    }
 
     /// Retains only the elements specified by the predicate. This will only mark the length of the vector
     /// and items after the first removed item as dirty.
@@ -70,45 +69,10 @@ pub trait VecStoreExt: private::Sealed {
     /// store.retain(|&x| x % 2 == 0);
     /// assert_eq!(store.len(), 2);
     /// ```
-    fn retain(&mut self, f: impl FnMut(&Self::Item) -> bool);
-}
-
-impl<T, W> private::Sealed for Store<Vec<T>, W>
-where
-    W: Writable<Target = Vec<T>> + Copy + 'static,
-    T: 'static,
-{
-}
-
-impl<W: Writable<Target = Vec<T>> + Copy + 'static, T: 'static> VecStoreExt for Store<Vec<T>, W> {
-    type Item = T;
-
-    fn push(&mut self, value: Self::Item) {
-        self.selector().mark_dirty_shallow();
-        self.selector().write.write_unchecked().push(value);
-    }
-
-    fn remove(&mut self, index: usize) -> Self::Item {
-        self.selector().mark_dirty_shallow();
-        self.selector().mark_dirty_at_and_after_index(index);
-        self.selector().write.write_unchecked().remove(index)
-    }
-
-    fn insert(&mut self, index: usize, value: Self::Item) {
-        self.selector().mark_dirty_shallow();
-        self.selector().mark_dirty_at_and_after_index(index);
-        self.selector().write.write_unchecked().insert(index, value);
-    }
-
-    fn clear(&mut self) {
-        self.selector().mark_dirty();
-        self.selector().write.write_unchecked().clear();
-    }
-
-    fn retain(&mut self, mut f: impl FnMut(&Self::Item) -> bool) {
+    pub fn retain(&mut self, mut f: impl FnMut(&T) -> bool) {
         let mut index = 0;
         let mut first_removed_index = None;
-        self.selector().write.write_unchecked().retain(|item| {
+        self.selector().write_untracked().retain(|item| {
             let keep = f(item);
             if !keep {
                 first_removed_index = first_removed_index.or(Some(index));
